added po files for all applets + updated the build script to build translations + updated conf files to match the common sentences of the C plug-ins

+++ make_packages.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 # grab the list of applets to package from the list.conf.
 list=`sed -n "/^\[.*\]/p" list.conf | tr -d "[]"`
@@ -10,13 +10,15 @@ else
 	mkdir FTP
 	cp list.conf FTP
 	for f in $list; do
+		test ! -e "$f/auto-load.conf" && continue
+		
 		echo "make $f"
 		# remove unwanted files.
 		rm -f "$f/*.pyc" "$f/*~"
 		
 		# check that the version in noth .conf are identical.
 		version1=`grep "^version *= *" "$f/auto-load.conf" | sed "s/.*= *//g"`
-		version2=`head -1 "$f/$f.conf" | sed "s/.*;//g"`
+		version2=`head -1 "$f/$f.conf" | tr -d "#"`
 		if test "$version1" != "$version2"; then
 			echo "  Warning: versions mismatch for $f ($version1/$version2)"
 		fi
@@ -28,6 +30,24 @@ else
 		mkdir "FTP/$f"
 		mv "$f.tar.gz" "FTP/$f"
 	done;
+	
+	# build language tree
+	f="locale"
+	name="cairo-dock-plugins-extra"
+	echo "make $f"
+	mkdir $f
+	
+	for p in po/*.po; do
+		pofile=${p:3}
+		lang=${pofile/.po/}  # filtrer "en" ?...
+		mkdir -p $f/${lang}/LC_MESSAGES
+		msgfmt -o $f/${lang}/LC_MESSAGES/${name}.mo $p
+	done;
+	
+	tar cfz "$f.tar.gz" $f
+	mkdir "FTP/$f"
+	mv "$f.tar.gz" "FTP/$f"
+	rm -rf $f
 fi
 
 exit 0
